requirements.txt
fastapi~=0.114.1
uvicorn
sqlalchemy~=2.0.36
pydantic~=2.9.1
passlib~=1.7.4
bcrypt
python-jose
pytest~=8.3.3
pydantic-settings
psycopg2-binary
python-multipart
requests
openai
jose~=1.0.0
together~=1.3.5
pandas~=2.2.3
langchain~=0.3.8